restrict equipment for user

+++ src/routers/equipment.py
@@ -1,17 +1,22 @@
-from fastapi import APIRouter, Depends
+from fastapi import APIRouter, Depends, HTTPException
 from sqlalchemy.orm import Session
 
 from D2Shared.shared.schemas.equipment import ReadEquipmentSchema, UpdateEquipmentSchema
 from src.database import session_local
-from src.models.rune import Equipment, Line
+from src.models.equipment import Equipment
+from src.models.rune import Line
+from src.models.user import User
 from src.queries.utils import get_or_create
+from src.security.auth import login
 
-router = APIRouter(prefix="/equipment")
+router = APIRouter(prefix="/equipment", dependencies=[Depends(login)])
 
 
 @router.post("/", response_model=ReadEquipmentSchema)
 def create_equipment(
-    equipment_datas: UpdateEquipmentSchema, session: Session = Depends(session_local)
+    equipment_datas: UpdateEquipmentSchema,
+    session: Session = Depends(session_local),
+    user: User = Depends(login),
 ):
     equipment = Equipment(label=equipment_datas.label)
     session.add(equipment)
@@ -20,7 +25,12 @@ def create_equipment(
     line_instances: list[Line] = []
     for line_schema in equipment_datas.lines:
         line = get_or_create(
-            session, Line, False, **line_schema.model_dump(), equipment_id=equipment.id
+            session,
+            Line,
+            False,
+            **line_schema.model_dump(),
+            equipment_id=equipment.id,
+            user_id=user.id,
         )[0]
         line_instances.append(line)
 
@@ -33,13 +43,21 @@ def update_equipment(
     equipment_id: int,
     equipment_datas: UpdateEquipmentSchema,
     session: Session = Depends(session_local),
+    user: User = Depends(login),
 ):
     equipment = session.get_one(Equipment, equipment_id)
+    if equipment.user_id != user.id:
+        raise HTTPException(403, "Can't update equipment of other users")
 
     line_instances: list[Line] = []
     for line_schema in equipment_datas.lines:
         line = get_or_create(
-            session, Line, False, **line_schema.model_dump(), equipment_id=equipment_id
+            session,
+            Line,
+            False,
+            **line_schema.model_dump(),
+            equipment_id=equipment_id,
+            user_id=user.id,
         )[0]
         line_instances.append(line)
 
@@ -49,13 +67,22 @@ def update_equipment(
 
 
 @router.delete("/{equipment_id}")
-def delete_equipment(equipment_id: int, session: Session = Depends(session_local)):
-    equipment = session.get(Equipment, equipment_id)
+def delete_equipment(
+    equipment_id: int,
+    session: Session = Depends(session_local),
+    user: User = Depends(login),
+):
+    equipment = session.get_one(Equipment, equipment_id)
+    if equipment.user_id != user.id:
+        raise HTTPException(403, "Can't delete equipment of other users")
     session.delete(equipment)
     session.commit()
 
 
 @router.get("/", response_model=list[ReadEquipmentSchema])
-def get_equipments(session: Session = Depends(session_local)):
-    equipments = session.query(Equipment).all()
+def get_equipments(
+    session: Session = Depends(session_local),
+    user: User = Depends(login),
+):
+    equipments = session.query(Equipment).filter(Equipment.user_id == user.id).all()
     return equipments
